+++ Views/ContentView.swift
@@ -205,3 +205,10 @@ struct ContentView: View {
         }
     }
 }
+
+#if DEBUG
+#Preview {
+    ContentView()
+        .modelContainer(PreviewSupport.makeContainer())
+}
+#endif